+++ typescript/src/services/deposits/index.ts
@@ -1,3 +1,4 @@
 export * from "./deposit"
 export * from "./deposits-service"
 export * from "./funding"
+export * from "./refund"
