+++ typescript/test/hex.test.ts
@@ -1,5 +1,5 @@
 import { assert } from "chai"
-import { Hex } from "../src/lib/utils"
+import { Hex } from "../src"
 
 describe("Hex", () => {
   const stringUnprefixed =
